init script works

+++ DDL/tables/clients.sql
@@ -2,7 +2,7 @@ CREATE TABLE clients (
   user_id        BIGINT PRIMARY KEY
                  REFERENCES users(user_id) ON DELETE RESTRICT,
   -- add domain specific data we want to track
-  age            TINYINT,
+  age            SMALLINT,
   marital_status TEXT,
   yearly_income  INT,
   created_at     TIMESTAMPTZ NOT NULL DEFAULT now(),
